+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.14.0] - June 25, 2023
+
+### Added
+
+- Add the book editions route - [(ISSUE)](https://github.com/nesaku/BiblioReads/issues/10)
+- Added "View More Editions" link to the book page
+
+### Fixed
+
+- Fix open source inconsistency in the README - [(ISSUE)](https://codeberg.org/nesaku/BiblioReads/issues/65)
+
 ## [2.13.1] - May 13, 2023
 
 ### Added